server/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
langchain
llama-index
python-dotenv
openai